server/requirements-ocr.txt
paddleocr==2.7.3
paddlepaddle==2.6.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
